requirements.txt
aiofiles==24.1.0
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
Django==5.1.2
django-jazzmin==3.0.1
django-solo==2.4.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.1.3
Pillow==9.5.0
proglog==0.1.10
psycopg2==2.9.10
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
requests==2.32.3
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.0
tzlocal==5.2
urllib3==2.2.3
gunicorn
